+++ src/@types/parser.d.ts
@@ -1,4 +1,4 @@
-import type { Cst } from '..';
+import type { Cst } from '../index.js';
 
 declare module '*/parser.js' {
 	// FIXME: 型指定が効いていない