+++ src/interpreter/lib/std.ts
@@ -1,10 +1,10 @@
 /* eslint-disable no-empty-pattern */
 import { v4 as uuid } from 'uuid';
 import seedrandom from 'seedrandom';
-import { NUM, STR, FN_NATIVE, FALSE, TRUE, ARR, NULL, BOOL, OBJ } from '../value';
-import { assertNumber, assertString, assertBoolean, valToJs, jsToVal, assertFunction, assertObject, eq, expectAny, assertArray, reprValue } from '../util';
-import { RuntimeError } from '../../error';
-import type { Value } from '../value';
+import { NUM, STR, FN_NATIVE, FALSE, TRUE, ARR, NULL, BOOL, OBJ, ERROR } from '../value.js';
+import { assertNumber, assertString, assertBoolean, valToJs, jsToVal, assertFunction, assertObject, eq, expectAny, assertArray, reprValue } from '../util.js';
+import { RuntimeError } from '../../error.js';
+import type { Value } from '../value.js';
 
 export const std: Record<string, Value> = {
 	'help': STR('SEE: https://github.com/syuilo/aiscript/blob/master/docs/get-started.md'),
@@ -152,7 +152,11 @@ export const std: Record<string, Value> = {
 
 	'Json:parse': FN_NATIVE(([json]) => {
 		assertString(json);
-		return jsToVal(JSON.parse(json.value));
+		try {
+			return jsToVal(JSON.parse(json.value));
+		} catch (e) {
+			return ERROR('not_json');
+		}
 	}),
 
 	'Json:parsable': FN_NATIVE(([str]) => {
@@ -449,6 +453,7 @@ export const std: Record<string, Value> = {
 			return NUM(1);
 		}
 	}),
+
 	'Str:gt': FN_NATIVE(([a, b]) => {
 		assertString(a);
 		assertString(b);
@@ -459,6 +464,12 @@ export const std: Record<string, Value> = {
 		} else {
 			return NUM(1);
 		}
+	}),
+
+	'Str:from_codepoint': FN_NATIVE(([codePoint]) => {
+		assertNumber(codePoint);
+
+		return STR(String.fromCodePoint(codePoint.value));
 	}),
 	//#endregion