+++ src/interpreter/primitive-props.ts
@@ -1,9 +1,14 @@
 import { substring, length, indexOf, toArray } from 'stringz';
-import { assertArray, assertBoolean, assertFunction, assertNumber, assertString, expectAny } from './util';
-import { ARR, FALSE, FN_NATIVE, NULL, NUM, STR, TRUE } from './value';
-import type { Value, VArr, VFn, VNum, VStr } from './value';
+import { RuntimeError } from '../error.js';
+import { assertArray, assertBoolean, assertFunction, assertNumber, assertString, expectAny } from './util.js';
+import { ARR, FALSE, FN_NATIVE, NULL, NUM, STR, TRUE } from './value.js';
+import type { Value, VArr, VFn, VNum, VStr, VError } from './value.js';
 
-export const PRIMITIVE_PROPS = {
+type VWithPP = VNum|VStr|VArr|VError;
+
+const PRIMITIVE_PROPS: {
+	[key in VWithPP['type']]: { [key: string]: (target: Value) => Value }
+} = {
 	num: {
 		to_str: (target: VNum): VFn => FN_NATIVE(async (_, _opts) => {
 			return STR(target.value.toString());
@@ -67,6 +72,14 @@ export const PRIMITIVE_PROPS = {
 			const chars = toArray(target.value);
 			const char = chars[i.value];
 			return char ? STR(char) : NULL;
+		}),
+
+		codepoint_at: (target: VStr): VFn => FN_NATIVE(([i], _) => {
+			assertNumber(i);
+
+			const res = target.value.charCodeAt(i.value);
+
+			return Number.isNaN(res) ? NULL : NUM(res);
 		}),
 	},
 
@@ -208,4 +221,23 @@ export const PRIMITIVE_PROPS = {
 			return target;
 		}),
 	},
+
+	error: {
+		name: (target: VError): VStr => STR(target.value), 
+
+		info: (target: VError): Value => target.info ?? NULL,
+	},
 } as const;
+
+export function getPrimProp(target: Value, name: string): Value {
+	if (Object.hasOwn(PRIMITIVE_PROPS, target.type)) {
+		const props = PRIMITIVE_PROPS[target.type as VWithPP['type']];
+		if (Object.hasOwn(props, name)) {
+			return props[name]!(target);
+		} else {
+			throw new RuntimeError(`No such prop (${name}) in ${target.type}.`);
+		}
+	} else {
+		throw new RuntimeError(`Cannot read prop of ${target.type}. (reading ${name})`);
+	}
+}